restorino V Final/requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.7
SQLAlchemy==2.0.20
email-validator==2.0.0
Pillow==10.0.0
python-dotenv==1.0.0
